Fix: CraftCanvas
여러 작업이 존재할 때, 생산, 획득에 따라, 보관량을 고려해서 아이템을 만들지 만들지 않을지를 결정하도록 수정

+++ Assets/EvolveThisMatch/Script/Lobby/Craft/DepartmentTemplate.cs
@@ -79,7 +79,7 @@ namespace EvolveThisMatch.Lobby
         public int maxUnits;
 
         [Tooltip("보관 가능한 총 무게")]
-        public float storageWeight;
+        public int storageWeight;
 
         [Tooltip("생산 속도 배율")]
         public float speed;
